rename(#12): 위젯클래스 폴더구조 변경 및 적용
